+++ src/client/widgets/basics/button.js
@@ -279,6 +279,13 @@ class Button extends Widget {
                 if (options.fromExternal) {
                     this.state = newstate
                     this.decoupledValue.setValue(newstate, {sync: true})
+                } else if (!options.local && this.getProp('mode') === 'push') {
+                    if (newstate === 1 && !this.touchActive) {
+                        this.touchActive = 1
+                        setTimeout(()=>{this.touchActive = 0})
+                    } else {
+                        this.touchActive = 0
+                    }
                 }
 
             } else {